Make bitmap_scan a feature vs. count freed entries
Dropping a pool needs to check if there are any entries are still in use. This asserts (in
debug mode) or leaks (in release mode). This can be done by and expensive scan of the bitmap
or by adding a counter. The later adds overhead to each allocation/free. Thus when dropping
pools is rare the bitmap scan should lead to better overall performance.